Add marked swipes to cleanup

+++ domain/src/main/java/ru/kima/intelligentchat/domain/common/useCase/CleanUpUseCase.kt
@@ -1,12 +1,21 @@
 package ru.kima.intelligentchat.domain.common.useCase
 
 import ru.kima.intelligentchat.domain.chat.repository.MessageRepository
+import ru.kima.intelligentchat.domain.chat.repository.SwipeRepository
 
 class CleanUpUseCase(
-    private val messageRepository: MessageRepository
+    private val messageRepository: MessageRepository,
+    private val swipeRepository: SwipeRepository
 ) {
     suspend operator fun invoke() {
         val markedMessages = messageRepository.getMarkedMessages()
-        messageRepository.deleteMessages(markedMessages)
+        if (markedMessages.isNotEmpty()) {
+            messageRepository.deleteMessages(markedMessages)
+        }
+
+        val markedSwipes = swipeRepository.getMarkedSwipes()
+        if (markedSwipes.isNotEmpty()) {
+            swipeRepository.deleteSwipes(markedSwipes)
+        }
     }
 }
